backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
pgvector==0.2.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Cache and async
redis==5.0.1
aiohttp==3.9.1
aiofiles==23.2.1

# HTTP client
httpx==0.25.2

# OCR integration (Baidu)
baidu-aip==4.16.13

# OSS integration (Aliyun)
oss2==2.18.4

# AI/LLM
openai==1.3.7
chromadb==0.4.22

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
email-validator==2.1.0
